django_adminFX-master/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==5.2.1
django-admin-tools==0.9.3
django-environ==0.12.0
django-simple-history==3.8.0
mysqlclient==2.2.7
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
